the_tuesday_club_server/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.2
django-ninja==1.3.0
pillow==10.4.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
dj-database-url== 2.3.0
brotli==1.1.0 
whitenoise==6.8.2
gunicorn==23.0.0 
h11==0.14.0 
packaging==24.2 
uvicorn==0.32.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
dnspython==2.7.0 
email-validator==2.2.0